requirements.txt
gitpython>=3.1.40
pyyaml>=6.0
python-dateutil>=2.8.2
openai>=1.3.0
anthropic>=0.7.0
jinja2>=3.1.2
